Renamed links and folders
