+++ src/App.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Route } from 'react-router-dom';
 import './App.css';
-import Debit from './components/Splits/Debit';
-import Credit from './components/Splits/Credit';
+import Debt from './components/Splits/Debt';
+import MyBills from './components/Splits/MyBills';
 import Login from './components/Login/Login';
 import Signup from './components/Signup/Signup';
 import withAuthCheck from './AuthCheck';
@@ -15,8 +15,8 @@ function App() {
       <Route exact path="/" component={Login} />
       <Route path="/signup" component={Signup} />
       <Route path='/dashboard' render={props => withAuthCheck(Dashboard, props)} />
-      <Route path='/my_credit' render={props => withAuthCheck(Credit, props)} />
-      <Route path='/my_debit' render={props => withAuthCheck(Debit, props)} />
+      <Route path='/my_bills' render={props => withAuthCheck(MyBills, props)} />
+      <Route path='/my_debt' render={props => withAuthCheck(Debt, props)} />
       <Route path='/create_bill' render={props => withAuthCheck(CreateBill, props)} />
     </div>
   );
